Provisional work on OnlineOthelloGames. Move OnlineOthelloGame into App.js, it's going to talk to sockets - I'll worry about precise naming/file convention later. Lots of work to discard/index.js and App.js to link the othello sessions & workings. I think the next commit might formalise a verbose mode in discard because I'm putting logging statements everywhere.
As of this commit - players are joining a session and being assigned roles successfully.
I had a bit of trouble because I was creating the socket in componentDidMount() but react seems to create and throwaway components multiple times when starting - this may be related to live-reloading/dev servers. Putting the appropriate cleanup in componentWillUnmount solved most of this.

Worth noting also that socket.io connections don't automatically disconnect! (I'm too used to c++'s class-based object-cleanup semantics here)